Rename boundingbox class
