+++ core/src/org/noble/helium/world/AxisOrientedBoundingBox.java
@@ -1,17 +1,17 @@
-package org.noble.helium.logic;
+package org.noble.helium.world;
 
 import com.badlogic.gdx.math.Matrix4;
 import com.badlogic.gdx.math.Vector3;
 import org.noble.helium.math.Dimensions3;
 import org.noble.helium.math.EulerAngles;
 
-public class OrientedBoundingBox {
+public class AxisOrientedBoundingBox {
   private final Vector3 center; // Floating-point center
   private final Vector3 halfSizes; // Half-width, half-height, half-depth
   private final Matrix4 rotationMatrix; // 4x4 Rotation matrix
   private final Vector3[] corners; // Transformed corner points
 
-  public OrientedBoundingBox(Vector3 position, Dimensions3 dimensions, EulerAngles angles) {
+  public AxisOrientedBoundingBox(Vector3 position, Dimensions3 dimensions, EulerAngles angles) {
     this.center = position;
     this.halfSizes = new Vector3(dimensions.getWidth() / 2f, dimensions.getHeight() / 2f, dimensions.getDepth() / 2f);
     this.rotationMatrix = createRotationMatrix(angles.getYaw(), angles.getPitch(), angles.getRoll());
@@ -50,7 +50,7 @@ public class OrientedBoundingBox {
     return localCorners;
   }
 
-  public boolean isColliding(OrientedBoundingBox other) {
+  public boolean isColliding(AxisOrientedBoundingBox other) {
     Vector3[] axes = getSeparatingAxes(other);
 
     for (Vector3 axis : axes) {
@@ -61,7 +61,7 @@ public class OrientedBoundingBox {
     return true; // No separating axis found → Collision
   }
 
-  private Vector3[] getSeparatingAxes(OrientedBoundingBox other) {
+  private Vector3[] getSeparatingAxes(AxisOrientedBoundingBox other) {
     Vector3[] axesA = getLocalAxes();
     Vector3[] axesB = other.getLocalAxes();
     Vector3[] axes = new Vector3[15]; // Ensure all axes are assigned
@@ -87,7 +87,7 @@ public class OrientedBoundingBox {
     return axes;
   }
 
-  private boolean overlapsOnAxis(Vector3 axis, OrientedBoundingBox other) {
+  private boolean overlapsOnAxis(Vector3 axis, AxisOrientedBoundingBox other) {
     float minA = Float.POSITIVE_INFINITY, maxA = Float.NEGATIVE_INFINITY;
     float minB = Float.POSITIVE_INFINITY, maxB = Float.NEGATIVE_INFINITY;
